sql_bot/requirements.txt
python-dotenv
langchain
langchain-community
openai
pymysql
sqlalchemy
